requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
resend>=0.7.0
pyyaml>=6.0.1
playwright>=1.40.0

